mcp-client/requirements.txt
mcp>=1.1.0
httpx>=0.27.0
google-genai>=0.1.0
streamlit>=1.51.0